Avoid pgoff overflow in remap_file_pages
commit 5ec1055aa5632dd7a8283cdb5fa9be3c535eaa06 upstream

Thomas Pollet noticed that the remap_file_pages() system call in
fremap.c has a potential overflow in the first part of the if statement
below, which could cause it to process bogus input parameters.
Specifically the pgoff + size parameters could be wrap thereby
preventing the system call from failing when it should.

Reported-by: Thomas Pollet <thomas.pollet@gmail.com>
Signed-off-by: Larry Woodman <lwoodman@redhat.com>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Cherry-picked for: tip/rt/2.6.33
Signed-off-by: John Kacur <jkacur@redhat.com>