x86: Check tsc available/disabled in the delayed init function commit a8760eca6cf60ed303ad494ef45901f63165d2c8 from linux-tip
Bugzilla: 660306

The delayed TSC init function does not check whether the system has no
TSC or TSC is disabled at the kernel command line, which results in a
crash in the work queue based extended calibration due to division by
zero because the basic calibration never happened.

Add the missing checks and do not touch TSC when not available or
disabled.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Cc: John Stultz <johnstul@us.ibm.com>
Signed-off-by: Luis Claudio R. Gonçalves <lgoncalv@redhat.com>