BZ677269 xfs: prevent leaking uninitialized stack memory in FSGEOMETRY_V1
xfs: prevent leaking uninitialized stack memory in FSGEOMETRY_V1

Bugzilla: 677269

The FSGEOMETRY_V1 ioctl (and its compat equivalent) calls out to
xfs_fs_geometry() with a version number of 3.  This code path does not
fill in the logsunit member of the passed xfs_fsop_geom_t, leading to
the leaking of four bytes of uninitialized stack data to potentially
unprivileged callers.  Since all other members are filled in all code
paths and there are no padding bytes in this structure, it's safe to
avoid an expensive memset() in favor of just clearing this one field.

Signed-off-by: Dan Rosenberg <drosenberg@vsecurity.com>
Signed-off-by: Luis Claudio R. Gonçalves <lgoncalv@redhat.com>
Signed-off-by: John Kacur <jkacur@redhat.com>