ksoftirqd should not do a trylock loop with -rt
Last night after disabling rcu stall and hung task, I just let the
system run. This morning the system was unresponsive, but it still had
sysrq-t working. Doing a full task dump and showing all the debug info
that I added, Thomas Gleixner was able to see that ksoftirqd had
preempted sshd, which was bound to a CPU with migration disabled.
Another task that was blocked on a lock that sshd had, had the
qdisc_lock. ksoftirqd was in net_tx_action() doing a loop where it did a
spin_trylock() on the qdisc_lock and if it failed, it added itself back
to trigger another softirq and continued. But by adding itself back,
ksoftirqd ran again, and as ksoftirqd is an RT task, it never let sshd
run. The problem was that the qdisc_lock was held by a task that was
blocked on a lock held by sshd.

This is not an issue in mainline as the locks in question are real spin
locks that disable interrupts or softirqs when taken, and the issue of
the preemption would not have happened. Regardless, the code in
net_tx_action() is a bit silly, and Thomas will be talking with David
Miller on changing it.

For now, we found a fix. For PREEMPT_RT_FULL, we just do a spin_lock()
instead of a spin_trylock() on the qdisc lock, and my testing has shown
that it makes the system run much better. More importantly, it removes
the live lock that was causing the RCU stalls and hung tasks.

Solved-by: Thomas Gleixner <tglx@linuxtronix.de>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>